Remove additional information in list-all output
`xcodes list` prints the Xcode build number and if the listed version is
currently installed or selected. Unfortunately this information is space
separated which makes asdf thinks they're all different versions. This
causes the following:

    12.2 Release Candidate (12B5044c)
    12.2 (12B45b) (Installed)

to be listed as:

    12.2 Release Candidate (12B5044c)12.2 (12B45b) (Installed)

We considered separating all this information with dashes but in reality
it's superfluous. In almost all case the version number is enough
information to install the correct Xcode.

Luckily all the superfluous information always starts with a space
followed by the information surrounded with parentheses. This uses sed
to look for that pattern and replaces it with nothing. Leftover
whitespace is then replaced with '-'.

The end result is that the previous example is listed as:

    12.2-Release-Candidate
    12.2

This also changes the shebang from /bin/sh to /bin/bash. While we prefer
/bin/sh the while loop is easier to write in bash. Furthermore right now
asdf literally calls out to bash to run our list-all script:
https://github.com/asdf-vm/asdf/blob/456d8e36ca76b36b848453a63b54ffeb214bce7e/lib/commands/command-list-all.bash#L11